Test if downloading works
